[Refactor] 변경된 내용을 토대로 test 코드 수정 #203
