+++ src/test/java/com/yzgeneration/evc/chat/service/ChatServiceTest.java
@@ -7,12 +7,11 @@ import com.yzgeneration.evc.domain.chat.implement.ChatConnectionManager;
 import com.yzgeneration.evc.domain.chat.implement.ChatRoomManager;
 import com.yzgeneration.evc.domain.chat.implement.SessionAttributeAccessor;
 import com.yzgeneration.evc.domain.chat.infrastructure.ChatConnectionRepository;
-import com.yzgeneration.evc.domain.chat.infrastructure.ChatMemberRepository;
 import com.yzgeneration.evc.domain.chat.infrastructure.ChatMessageRepository;
 import com.yzgeneration.evc.domain.chat.model.ChatMember;
 import com.yzgeneration.evc.domain.chat.model.ChatMessage;
 import com.yzgeneration.evc.domain.chat.service.ChatService;
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import com.yzgeneration.evc.exception.CustomException;
 import com.yzgeneration.evc.exception.ErrorCode;
 import com.yzgeneration.evc.mock.chat.FakeChatConnectionRepository;
