+++ src/test/java/com/yzgeneration/evc/mock/delivery/SpyDeliveryRepository.java
@@ -5,7 +5,7 @@ import com.yzgeneration.evc.domain.delivery.dto.MobilityCreate;
 import com.yzgeneration.evc.domain.delivery.infrastructure.DeliveryRepository;
 import com.yzgeneration.evc.domain.delivery.model.Delivery;
 import com.yzgeneration.evc.domain.delivery.model.DeliveryView;
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import com.yzgeneration.evc.exception.CustomException;
 import com.yzgeneration.evc.exception.ErrorCode;
 import com.yzgeneration.evc.fixture.DeliveryFixture;
