+++ src/test/java/com/yzgeneration/evc/mock/image/FakeItemImageRepository.java
@@ -1,6 +1,6 @@
 package com.yzgeneration.evc.mock.image;
 
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import com.yzgeneration.evc.domain.image.model.ItemImage;
 import com.yzgeneration.evc.domain.image.service.port.ItemImageRepository;
